% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_gap_finding.R
\name{genome_gap_finding}
\alias{genome_gap_finding}
\title{Genome gap finding}
\usage{
genome_gap_finding(fasta, chromosomes, min_gap = 50, gap_symbol = "N")
}
\arguments{
\item{fasta}{Fasta file conting genome sequence}

\item{chromosomes}{Selected chromosomes}

\item{min_gap}{Minimum gap size}

\item{gap_symbol}{Gap symbol, default is 'N'}
}
\value{
A plot showing the gap regions
}
\description{
Genome gap finding
}
\examples{
plot <- genome_gap_finding("genome.fasta", c("2", "3"), min_gap=50, gap_symbol="N")
}
