% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paf_selected_coverage.R
\name{paf_selected_coverage}
\alias{paf_selected_coverage}
\title{Plot the coverage of a selected contig}
\usage{
paf_selected_coverage(
  data,
  chr = "Contig119",
  mapping_type = "P",
  target_region = NULL
)
}
\arguments{
\item{data}{alignment matrix}

\item{chr}{selected contig}

\item{mapping_type}{selected type, 'P' or 'S'}

\item{target_region}{selected target region}
}
\value{
a ggplot object
}
\description{
Plot the coverage of a selected contig
}
\examples{
# Example usage:
paf_selected_coverage(ali_selected, chr="chr2", mapping_type="P")
}
