% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genomic_coverage_and_annotations.R
\name{plot_genomic_coverage_and_annotations}
\alias{plot_genomic_coverage_and_annotations}
\title{Plot Genomic Coverage and Gene Annotations}
\usage{
plot_genomic_coverage_and_annotations(
  bam,
  gtf,
  chromosome,
  region_start,
  region_end,
  by = 500,
  smooth = FALSE,
  smooth_color = "red",
  bin_coverage = FALSE,
  bin_size = 500,
  coverage_color = "blue",
  annotation_color = "yellow"
)
}
\arguments{
\item{bam}{The path to the BAM file.}

\item{gtf}{The path to the GTF file.}

\item{chromosome}{The chromosome or contig name (e.g., "2R").}

\item{smooth}{Logical; if TRUE, applies a smooth line to the coverage plot.}

\item{smooth_color}{The color of the smooth line. Default is 'red'.}

\item{bin_coverage}{Logical; if TRUE, calculates average coverage per bin.}

\item{bin_size}{Numeric; bin size for averaging coverage, applicable if use_bin_coverage is TRUE.}

\item{coverage_color}{The color of the coverage plot. Default is 'blue'.}

\item{annotation_color}{The color of the gene annotations. Default is 'yellow'.}

\item{start}{The start position of the genomic region.}

\item{end}{The end position of the genomic region.}

\item{break}{The break size for the coverage plot.}
}
\value{
A combined ggplot object showing the coverage plot and gene annotations.
}
\description{
This function plots genomic coverage from a BAM file and gene annotations from a GTF file.
}
\examples{
# Example usage:
plot_genomic_coverage_and_annotations(
  bamfile = "path/to/bamfile.bam",
  gtf_file = "path/to/gtf_file.gtf",
  chromosome = "2R",
  visible_start = 13484000,
  visible_end = 13487000,
  use_smooth = FALSE,
  use_bin_coverage = FALSE,
  bin_size = 500
)
}
