% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paf.R
\name{read.paf}
\alias{read.paf}
\title{Read paf file into a dataframe}
\usage{
read.paf(file, header = FALSE, tag = TRUE)
}
\arguments{
\item{file}{paf file generated from minimap2}

\item{header}{whether the paf file has header}

\item{tag}{whether the paf file has tag}
}
\value{
paf data in a dataframe
}
\description{
Read paf file into a dataframe
}
\examples{
read.paf("test.paf", header = FALSE, tag = TRUE)
}
